Added two exceptions and body for editing a generic Post.

+++ generic-service-post/src/main/java/nl/tudelft/sem/genericservicepost/exceptions/GenericPostNotFoundException.java
@@ -5,5 +5,5 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(value = HttpStatus.NOT_FOUND, reason = "Generic post ID was not found!")
 public class GenericPostNotFoundException extends RuntimeException{
-    public static final long serialVersionUID = (long) 1;
+    public static final long serialVersionUID = 1L;
 }
