+++ generic-service-post/src/main/java/nl/tudelft/sem/genericservicepost/exceptions/InvalidEditException.java
@@ -5,5 +5,5 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(value = HttpStatus.NOT_MODIFIED, reason = "NetId can't be edited")
 public class InvalidEditException extends RuntimeException{
-    public static final long serialVersionUID = (long) 2;
+    public static final long serialVersionUID = 2L;
 }
